requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
discord.py==2.0.1
ffmpeg==1.4
idna==3.3
multidict==5.2.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
requests==2.28.1
typing_extensions==4.0.1
urllib3==1.26.12
yarl==1.7.2
youtube-dl==2021.12.17
